AlirezaMirzaei/Problem4_WebAPI/requirements.txt
fastapi
uvicorn
celery
redis
docker
sqlalchemy
pydantic
psycopg2-binary
python-dotenv